final_touch

+++ app/Http/Controllers/ListingController.php
@@ -34,18 +34,20 @@ class ListingController extends Controller
     {
         # dd($request->all());
         # dd($request->file('logo'));
+        # dd(auth()->user()->id);
         $formFields = $request->validate([
-           'title' => 'required',
-           'company' => ['required', Rule::unique('listings', 'company')],
-           'location' => 'required',
-           'website' => 'required',
-           'email' => ['required', 'email'],
-           'tags' => 'required',
-           'description' => 'required'
+            'title' => 'required',
+            'company' => ['required', Rule::unique('listings', 'company')],
+            'location' => 'required',
+            'website' => 'required',
+            'email' => ['required', 'email'],
+            'tags' => 'required',
+            'description' => 'required'
         ]);
         if ($request->hasFile('logo')){
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
+        $formFields['user_id'] = auth()->id();
         Listing::create($formFields);
         return redirect('/')->with('message', 'Listing Created Successfully!');
     }
@@ -54,12 +56,20 @@ class ListingController extends Controller
     public function edit($id)
     {
         $listing = Listing::find($id);
+        // Make sure logged-in user is owner
+        if ($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
         return view('listings.edit')->with('listing', $listing);
     }
 
     // update Listing data
     public function update(Request $request, Listing $id)
     {
+        // Make sure logged-in user is owner
+        if ($id->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
         # dd($request->all());
         $formFields = $request->validate([
             'title' => 'required',
@@ -80,12 +90,22 @@ class ListingController extends Controller
     // delete Listing data
     public function destroy(Listing $id)
     {
-//        dd($id->logo);
+        // Make sure logged-in user is owner
+        if ($id->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+        # dd($id->logo);
         if ($id->logo != 'no-image.png'){
-//            dd($id->logo);
+            # dd($id->logo);
             Storage::delete('public/' . $id->logo);
         }
         $id->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    // Manage Listings
+    public function manage()
+    {
+        return view('listings.manage')->with(['listings' => auth()->user()->listings()->get()]);
     }
 }
